Week-5/iitb-ug-rules-chatbot/requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
huggingface_hub
sentence-transformers
faiss-cpu
python-dotenv
